+++ javascript/webdriver/http/http.js
@@ -43,18 +43,15 @@ webdriver.http.Client = function() {
 
 
 /**
- * Sends a request to the server. If an error occurs while sending the request,
- * such as a failure to connect to the server, the provided callback will be
- * invoked with a non-null {@link Error} describing the error. Otherwise, when
- * the server's response has been received, the callback will be invoked with a
- * null Error and non-null {@link webdriver.http.Response} object.
+ * Sends a request to the server. The client will automatically follow any
+ * redirects returned by the server, fulfilling the returned promise with the
+ * final response.
  *
  * @param {!webdriver.http.Request} request The request to send.
- * @param {function(Error, !webdriver.http.Response=)} callback the function to
- *     invoke when the server's response is ready.
+ * @return {!webdriver.promise.Promise<!webdriver.http.Response>} A promise that
+ *     will be fulfilled with the server's response.
  */
-webdriver.http.Client.prototype.send = function(request, callback) {
-};
+webdriver.http.Client.prototype.send = function(request) {};
 
 
 
@@ -106,7 +103,7 @@ webdriver.http.Executor.prototype.defineCommand = function(
 
 
 /** @override */
-webdriver.http.Executor.prototype.execute = function(command, callback) {
+webdriver.http.Executor.prototype.execute = function(command) {
   var resource =
       this.customCommands_[command.getName()] ||
       webdriver.http.Executor.COMMAND_MAP_[command.getName()];
@@ -123,21 +120,12 @@ webdriver.http.Executor.prototype.execute = function(command, callback) {
     return '>>>\n' + request;
   });
 
-  this.client_.send(request, function(e, response) {
-    var responseObj;
-    if (!e) {
-      log.finer(function() {
-        return '<<<\n' + response;
-      });
-      try {
-        responseObj = webdriver.http.Executor.parseHttpResponse_(
-            /** @type {!webdriver.http.Response} */ (response));
-      } catch (ex) {
-        log.warning('Error parsing response', ex);
-        e = ex;
-      }
-    }
-    callback(e, responseObj);
+  return this.client_.send(request).then(function(response) {
+    log.finer(function() {
+      return '<<<\n' + response;
+    });
+    return webdriver.http.Executor.parseHttpResponse_(
+          /** @type {!webdriver.http.Response} */ (response));
   });
 };