+++ javascript/webdriver/test/http/xhrclient_test.js
@@ -77,14 +77,11 @@ function testXhrClient_whenUnableToSendARequest() {
   });
   control.$replayAll();
 
-  var callback;
-  new webdriver.http.XhrClient(URL).send(REQUEST,
-      callback = callbackHelper(function(error) {
-        assertNotNullNorUndefined(error);
-        assertEquals(1, arguments.length);
-      }));
-  callback.assertCalled();
-  control.$verifyAll();
+  return new webdriver.http.XhrClient(URL)
+      .send(REQUEST)
+      .then(fail, function() {
+        control.$verifyAll();
+      });
 }
 
 function testXhrClient_parsesResponseHeaders_windows() {
@@ -101,11 +98,9 @@ function testXhrClient_parsesResponseHeaders_windows() {
   ].join('\r\n'));
   control.$replayAll();
 
-  var callback;
-  new webdriver.http.XhrClient(URL).send(REQUEST,
-      callback = callbackHelper(function(e, response) {
-        assertNull(e);
-
+  return new webdriver.http.XhrClient(URL)
+      .send(REQUEST)
+      .then(function(response) {
         assertEquals(200, response.status);
         assertEquals('', response.body);
 
@@ -115,9 +110,9 @@ function testXhrClient_parsesResponseHeaders_windows() {
           'e': 'f',
           'g': 'h'
         }, response.headers);
-      }));
-  callback.assertCalled();
-  control.$verifyAll();
+
+        control.$verifyAll();
+      });
 }
 
 function testXhrClient_parsesResponseHeaders_unix() {
@@ -134,10 +129,9 @@ function testXhrClient_parsesResponseHeaders_unix() {
   ].join('\n'));
   control.$replayAll();
 
-  var callback;
-  new webdriver.http.XhrClient(URL).send(REQUEST,
-      callback = callbackHelper(function(e, response) {
-        assertNull(e);
+  return new webdriver.http.XhrClient(URL)
+      .send(REQUEST)
+      .then(function(response) {
         assertEquals(200, response.status);
         assertEquals('', response.body);
 
@@ -147,9 +141,9 @@ function testXhrClient_parsesResponseHeaders_unix() {
           'e': 'f',
           'g': 'h'
         }, response.headers);
-      }));
-  callback.assertCalled();
-  control.$verifyAll();
+
+        control.$verifyAll();
+      });
 }
 
 function testXhrClient_handlesResponsesWithNoHeaders() {
@@ -161,17 +155,16 @@ function testXhrClient_handlesResponsesWithNoHeaders() {
   mockXhr.getAllResponseHeaders().$returns('');
   control.$replayAll();
 
-  var callback;
-  new webdriver.http.XhrClient(URL).send(REQUEST,
-      callback = callbackHelper(function(e, response) {
-        assertNull(e);
+  return new webdriver.http.XhrClient(URL)
+      .send(REQUEST)
+      .then(function(response) {
         assertEquals(200, response.status);
         assertEquals('', response.body);
 
         webdriver.test.testutil.assertObjectEquals({}, response.headers);
-      }));
-  callback.assertCalled();
-  control.$verifyAll();
+
+        control.$verifyAll();
+      });
 }
 
 function testXhrClient_stripsNullCharactersFromResponseBody() {
@@ -183,14 +176,12 @@ function testXhrClient_stripsNullCharactersFromResponseBody() {
   mockXhr.getAllResponseHeaders().$returns('');
   control.$replayAll();
 
-  var callback;
-  new webdriver.http.XhrClient(URL).send(REQUEST,
-      callback = callbackHelper(function(e, response) {
-        assertNull(e);
+  return new webdriver.http.XhrClient(URL)
+      .send(REQUEST)
+      .then(function(response) {
         assertEquals(200, response.status);
         assertEquals('foobar', response.body);
         webdriver.test.testutil.assertObjectEquals({}, response.headers);
-      }));
-  callback.assertCalled();
-  control.$verifyAll();
+        control.$verifyAll();
+      });
 }